trans-openapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt==2.8.0
cryptography==41.0.7
httpx==0.25.2
python-multipart==0.0.6
loguru==0.7.2
python-dotenv==1.0.0
redis==5.0.1
watchdog==3.0.0
pysqlite3>=0.5.0
